Refactor project screen (#173)
* Create component to choose project

* Write unit test

* Create component to create project, write unit tests

* Fix compilation problems

* Remove console log in DuplicateFinder

* Write more unit tests
