+++ src/components/ProjectScreen/CreateProject/index.tsx
@@ -1,5 +1,5 @@
 import CreateProject from "./CreateProjectComponent";
-import { StoreState } from "../../types";
+import { StoreState } from "../../../types";
 
 import { connect } from "react-redux";
 import { ThunkDispatch } from "redux-thunk";
@@ -8,8 +8,7 @@ import {
   asyncCreateProject,
   reset
 } from "./CreateProjectActions";
-import { Project } from "../../types/project";
-import { setCurrentProject, ProjectAction } from "../Project/ProjectActions";
+import { ProjectAction } from "../../Project/ProjectActions";
 
 function mapStateToProps(state: StoreState) {
   return {
@@ -26,9 +25,6 @@ export function mapDispatchToProps(
   return {
     asyncCreateProject: (name: string, languageData: File) => {
       dispatch(asyncCreateProject(name, languageData));
-    },
-    setCurrentProject: (project: Project) => {
-      dispatch(setCurrentProject(project));
     },
     reset: () => {
       dispatch(reset());
