+++ src/rootReducer.tsx
@@ -5,7 +5,7 @@ import { tempReducer } from "./components/Temp/TempReducer";
 import { goalsReducer } from "./components/GoalTimeline/GoalsReducer";
 import { localizeReducer } from "react-localize-redux";
 import { loginReducer } from "./components/Login/LoginReducer";
-import { createProjectReducer } from "./components/CreateProject/CreateProjectReducer";
+import { createProjectReducer } from "./components/ProjectScreen/CreateProject/CreateProjectReducer";
 import { characterInventoryReducer } from "./goals/CharInventoryCreation/CharacterInventoryReducer";
 import { goalSelectReducer } from "./components/GoalTimeline/GoalSwitcher/GoalSelectorScroll/GoalSelectorReducer";
 import { projectReducer } from "./components/Project/ProjectReducer";
