+++ src/types/index.tsx
@@ -1,7 +1,7 @@
 import { TempState } from "../components/Temp/TempReducer";
 import { LocalizeState } from "react-localize-redux";
 import { LoginState } from "../components/Login/LoginReducer";
-import { CreateProjectState } from "../components/CreateProject/CreateProjectReducer";
+import { CreateProjectState } from "../components/ProjectScreen/CreateProject/CreateProjectReducer";
 import { GoalsState, GoalSelectorState } from "./goals";
 import { CharacterInventoryState } from "../goals/CharInventoryCreation/CharacterInventoryReducer";
 import { Project } from "./project";
